requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# Optional dependencies untuk production
gunicorn==21.2.0
psycopg2-binary==2.9.7  # untuk PostgreSQL
PyJWT==2.8.0  # untuk API authentication

# Development dependencies (opsional)
pytest==7.4.2
pytest-cov==4.1.0
flask-testing==0.8.1
black==23.7.0
flake8==6.0.0